Ability to pause screen events & fixed some issues

+++ DP_demo/demo/android/app/src/main/java/com/dp/demo/ScreenEventsReceiver.java
@@ -20,6 +20,9 @@ public class ScreenEventsReceiver extends BroadcastReceiver {
 
     long ts = System.currentTimeMillis();
     File dir = context.getFilesDir();
+    //if a sentinel file exists, skip logging
+    File disabled = new File(dir, "screen-events.disabled");
+    if (disabled.exists()) return;
     File log = new File(dir, "screen-events.log");
 
     try (FileWriter writer = new FileWriter(log, true)) {
